Move ExportSelection to Export/common
